usb_updater: Cast chunk_len to size_t
chunk_len is int, payload_size and trasnfer_size are size_t. This
results in a wanrning when comparing with MIN. Fix by casting chunk_len
to size_t before comparing.

BUG=None
BRANCH=None
TEST=Build

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4311236

+++ extra/usb_updater/usb_updater2.c
@@ -501,7 +501,8 @@ static int transfer_block(struct usb_endpoint *uep,
 	for (transfer_size = 0; transfer_size < payload_size;) {
 		int chunk_size;
 
-		chunk_size = MIN(uep->chunk_len, payload_size - transfer_size);
+		chunk_size = MIN((size_t)uep->chunk_len,
+				 payload_size - transfer_size);
 		xfer(uep, transfer_data_ptr, chunk_size, NULL, 0, 0);
 		transfer_data_ptr += chunk_size;
 		transfer_size += chunk_size;
